CORE-14685: Migrate HSQLDB expression logic into the Converter.
