+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/query/parsing/VaultNamedQueryParserImpl.kt
@@ -2,15 +2,27 @@ package net.corda.ledger.persistence.query.parsing
 
 import net.corda.ledger.persistence.query.parsing.converters.VaultNamedQueryConverter
 import net.corda.ledger.persistence.query.parsing.expressions.VaultNamedQueryExpressionParser
+import net.corda.ledger.persistence.query.parsing.expressions.VaultNamedQueryExpressionParserImpl
 import net.corda.ledger.persistence.query.parsing.expressions.VaultNamedQueryExpressionValidator
+import net.corda.ledger.persistence.query.parsing.expressions.VaultNamedQueryExpressionValidatorImpl
+import org.osgi.service.component.annotations.Activate
+import org.osgi.service.component.annotations.Component
+import org.osgi.service.component.annotations.Reference
 
-abstract class AbstractVaultNamedQueryParserImpl(
+@Component(service = [ VaultNamedQueryParser::class ])
+class VaultNamedQueryParserImpl(
     private val expressionParser: VaultNamedQueryExpressionParser,
     private val expressionValidator: VaultNamedQueryExpressionValidator,
     private val converter: VaultNamedQueryConverter
 ) : VaultNamedQueryParser {
 
-    final override fun parseWhereJson(query: String): String {
+    @Activate
+    constructor(
+        @Reference
+        converter: VaultNamedQueryConverter
+    ) : this(VaultNamedQueryExpressionParserImpl(), VaultNamedQueryExpressionValidatorImpl(), converter)
+
+    override fun parseWhereJson(query: String): String {
         val expression = expressionParser.parse(query)
         expressionValidator.validateWhereJson(query, expression)
         val output = StringBuilder()
