--- components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/query/parsing/expressions/PostgresVaultNamedQueryExpressionParser.kt
@@ -1,9 +1,0 @@
-package net.corda.ledger.persistence.query.parsing.expressions
-
-import net.corda.ledger.persistence.query.parsing.Token
-
-class PostgresVaultNamedQueryExpressionParser : AbstractVaultNamedQueryExpressionParserImpl() {
-    override fun postProcess(outputTokens: MutableList<Token>): MutableList<Token> {
-        return outputTokens
-    }
-}
