+++ components/ledger/ledger-persistence/src/test/kotlin/net/corda/ledger/persistence/query/impl/parsing/PostgresProvider.kt
@@ -1,0 +1,9 @@
+package net.corda.ledger.persistence.query.impl.parsing
+
+import net.corda.orm.DatabaseType
+import net.corda.orm.DatabaseTypeProvider
+
+object PostgresProvider : DatabaseTypeProvider {
+    override val databaseType: DatabaseType
+        get() = DatabaseType.POSTGRES
+}
