+++ components/ledger/ledger-persistence/src/test/kotlin/net/corda/ledger/persistence/query/impl/parsing/PostgresVaultNamedQueryParserIntegrationTest.kt
@@ -1,8 +1,7 @@
 package net.corda.ledger.persistence.query.impl.parsing
 
-import net.corda.ledger.persistence.query.parsing.PostgresVaultNamedQueryParserImpl
-import net.corda.orm.DatabaseType
-import net.corda.orm.DatabaseTypeProvider
+import net.corda.ledger.persistence.query.parsing.VaultNamedQueryParserImpl
+import net.corda.ledger.persistence.query.parsing.converters.PostgresVaultNamedQueryConverter
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
@@ -14,11 +13,7 @@ import java.util.stream.Stream
 @Suppress("MaxLineLength")
 class PostgresVaultNamedQueryParserIntegrationTest {
 
-    private val databaseTypeProvider = object : DatabaseTypeProvider {
-        override val databaseType: DatabaseType
-            get() = DatabaseType.POSTGRES
-    }
-    private val vaultNamedQueryParser = PostgresVaultNamedQueryParserImpl(databaseTypeProvider)
+    private val vaultNamedQueryParser = VaultNamedQueryParserImpl(PostgresVaultNamedQueryConverter(PostgresProvider))
 
     private companion object {
         @JvmStatic
