+++ testing/ledger/ledger-hsqldb/src/test/kotlin/net/corda/testing/ledger/query/HsqldbProvider.kt
@@ -1,0 +1,9 @@
+package net.corda.testing.ledger.query
+
+import net.corda.orm.DatabaseType
+import net.corda.orm.DatabaseTypeProvider
+
+object HsqldbProvider : DatabaseTypeProvider {
+    override val databaseType: DatabaseType
+        get() = DatabaseType.HSQLDB
+}
